cms/pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = ""
authors = [
    {name = "AntoDono",email = "anthony20151128@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.14.0,<4.0.0)",
    "django-cors-headers (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
